Added "Resources" folder for textures and such
Also, the scrolling change is now an int. If you move the screen by non
integer values, it starts to get fuzzy. The drawn stuff is not aligned
with the screen pixels.

+++ GameViews.cpp
@@ -48,7 +48,7 @@ void GameViews::render()
 void GameViews::scroll(const sf::Time& clock)
 {
   //71 is left, 72 is right, 73 is up, 74 is downs
-  double change = clock.asMicroseconds()/100 * SCROLL_SPEED;
+  int change = clock.asMicroseconds()/100 * SCROLL_SPEED;
    if ( change > 50 )
      {
        change = 50;
